backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
boto3==1.34.14
botocore==1.34.14
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
requests==2.31.0
sounddevice
soundfile
